# NormalEstimationOMP: use dynamic scheduling for faster computation