@larshg 2023-08-03
Looks good :+1: 

I also read that there exist a **guided** as well as **runtime** (Set by env OMP_SCHEDULE variable) scheduling.

Btw. should we make directory with all these "compare" programs that gets created - I assume it would be nice to have such, when working with improving things.

Or do you use the output of ie. the already added google benchmarks and then do the math(speedup factor calculations) elsewhere?