@mvieth 2023-08-03
> I also read that there exist a **guided** as well as **runtime** (Set by env OMP_SCHEDULE variable) scheduling.

Yes, that's true. Technically, the guided schedule should have less overhead than the dynamic schedule. However, I read somewhere that the guided schedule is realized in a bad way in some OpenMP implementations, namely that the first chunk is too large and thus the work is again unbalanced between the threads. If I remember correctly, I tested the guided schedule some time ago and it was worse than the dynamic schedule for the normal estimation.

> Btw. should we make directory with all these "compare" programs that gets created - I assume it would be nice to have such, when working with improving things.
> 
> Or do you use the output of ie. the already added google benchmarks and then do the math(speedup factor calculations) elsewhere?

I wrote a quick Python script (see below) that reads from a json file, created by the google benchmark, and computes the speedup overview. But I don't think it is nice enough to put it into the repo permanently. I did however extend our google benchmark for the normal estimation, maybe I can make a pull request to add that sometime.

```python
#!/usr/bin/env python3
import json
import sys
with open(sys.argv[1]) as json_data:
    data = json.load(json_data)
average_speedup = 0
average_parallelization = 0
for dataset in ["mug", "milk"]:
    for typ in ["organized", "unorganized"]:
        search = "radius"
        for param in [10, 20]:
            time_w_omp = 1
            time_wo_omp = 1
            time_w_omp_cpu = 1
            for benchmark in data["benchmarks"]:
                if benchmark["name"] == "BM_NormalEstimation_" + dataset + "_" + typ + "_radius/" + str(param) + "/iterations:5/repeats:3_mean":
                    time_wo_omp = benchmark["real_time"]
                if benchmark["name"] == "BM_NormalEstimationOMP_" + dataset + "_" + typ + "_radius/" + str(param) + "/iterations:10/repeats:3/process_time/real_time_mean":
                    time_w_omp = benchmark["real_time"]
                    time_w_omp_cpu = benchmark["cpu_time"]
            print(dataset, typ, search, param, int(time_wo_omp+0.5), "/", int(time_w_omp+0.5), time_wo_omp/time_w_omp, time_w_omp_cpu/time_w_omp)
            average_speedup += time_wo_omp/time_w_omp
            average_parallelization += time_w_omp_cpu/time_w_omp
        search = "nearestk"
        for param in [50, 100]:
            time_w_omp = 1
            time_wo_omp = 1
            time_w_omp_cpu = 1
            for benchmark in data["benchmarks"]:
                if benchmark["name"] == "BM_NormalEstimation_" + dataset + "_" + typ + "_nearest_k/" + str(param) + "/iterations:5/repeats:3_mean":
                    time_wo_omp = benchmark["real_time"]
                if benchmark["name"] == "BM_NormalEstimationOMP_" + dataset + "_" + typ + "_nearest_k/" + str(param) + "/iterations:10/repeats:3/process_time/real_time_mean":
                    time_w_omp = benchmark["real_time"]
                    time_w_omp_cpu = benchmark["cpu_time"]
            print(dataset, typ, search, param, int(time_wo_omp+0.5), "/", int(time_w_omp+0.5), time_wo_omp/time_w_omp, time_w_omp_cpu/time_w_omp)
            average_speedup += time_wo_omp/time_w_omp
            average_parallelization += time_w_omp_cpu/time_w_omp
print("average speedup=", average_speedup/16)
print("average parallelization=", average_parallelization/16)
```